fix 'suppress_unshortopts' hack
Passing an Opt with short;

  SimpleOpts.get(foo: Opt.new(short: ?a)

breaks. (Except if the option name starts
with 'h', due to interference with --help
handling. short's value beiing nil also
triggers.)

I don't know how it supposed to work
up till now, the code in OptionParser
that is called here did not change in
the last 17 years.

Guess we'd be better off with tests...