add :choices keyword option to SimpleOpts::Opt
It takes a Set or a Hash. If given, set members / hash keys
will be accepted as valid values for the option, and in the
resultant record, the option name will be bound to the chosen
member / value for the chosen key. The option will be required
unless a default is given.

If the set / hash is directly passed to the SimpleOpts builder,
it spawns an Opt instance with this object as :choices and first
member / key as default.

Avao;ab;e choices are indcated in help output and in the error
message that's presented when the given value is not among the
choices.