allow Proc as value for :choices
Proc is limited compared to Set / Hash in following aspects: it's not
possible to derive a default choice and the list of available choices.

It neither responds to #key?, so instead of that we can just check for
nullity of result on key candidate (thus behavior on Hash suffers a slight
alteration).

Otherwise it presents the same interface as Set / Hash so we can support
for it for free. (The utility of it is not seen yet.)